pyproject.toml
[project]
name = "wikipediaparser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
